backend/requirements.txt
cocoindex[embeddings]>=0.3.20
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
psycopg[binary]>=3.0.0
psycopg-pool>=3.0.0
pgvector>=0.3.0
anthropic>=0.39.0
pydantic>=2.0.0
PyGithub>=2.0.0
